attribute macros for methods tracing.
