+++ .github/workflows/rust.yml
@@ -44,6 +44,6 @@ jobs:
     - name: Build
       run: cargo build --verbose
     - name: Run tests
-      run: cargo test --verbose -- --nocapture
+      run: cargo test --verbose --no-fail-fast 
     - name: Run benchmarks
       run: cargo bench
