+++ iroha_network/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+iroha_derive = { path = "../iroha_macro/iroha_derive" }
 async-std = { version = "1.5.0", features = ["attributes"] }
 futures = "0.3.4"
